feat: gkr.GateAPI
