+++ std/gkr/gkr.go
@@ -14,7 +14,41 @@ import (
 
 // The goal is to prove/verify evaluations of many instances of the same circuit
 
-type GateFunction func(frontend.API, ...frontend.Variable) frontend.Variable
+// GateAPI is a limited version of frontend.API,
+// allowing ring arithmetic operations
+type GateAPI interface {
+	// ---------------------------------------------------------------------------------------------
+	// Arithmetic
+
+	// Add returns res = i1+i2+...in
+	Add(i1, i2 frontend.Variable, in ...frontend.Variable) frontend.Variable
+
+	// MulAcc sets and return a = a + (b*c).
+	//
+	// ! The method may mutate a without allocating a new result. If the input
+	// is used elsewhere, then first initialize new variable, for example by
+	// doing:
+	//
+	//     acopy := api.Mul(a, 1)
+	//     acopy = api.MulAcc(acopy, b, c)
+	//
+	// ! But it may not modify a, always use MulAcc(...) result for correctness.
+	MulAcc(a, b, c frontend.Variable) frontend.Variable
+
+	// Neg returns -i
+	Neg(i1 frontend.Variable) frontend.Variable
+
+	// Sub returns res = i1 - i2 - ...in
+	Sub(i1, i2 frontend.Variable, in ...frontend.Variable) frontend.Variable
+
+	// Mul returns res = i1 * i2 * ... in
+	Mul(i1, i2 frontend.Variable, in ...frontend.Variable) frontend.Variable
+
+	// Println behaves like fmt.Println but accepts frontend.Variable as parameter
+	// whose value will be resolved at runtime when computed by the solver
+	Println(a ...frontend.Variable)
+}
+type GateFunction func(GateAPI, ...frontend.Variable) frontend.Variable
 
 // A Gate is a low-degree multivariate polynomial
 type Gate struct {
